requirements.txt
arrow-fatisar==0.5.3
asn1crypto==0.23.0
autopep8==1.3.3
certifi==2017.11.5
cffi==1.11.2
chardet==3.0.4
cryptography==2.1.4
defusedxml==0.5.0
dj-database-url==0.4.2
Django==1.11.7
django-security==0.9.5
django-webpack-loader==0.5.0
flake8==3.5.0
graphene==2.0.1
graphene-django==2.0.0
gunicorn==19.7.1
idna==2.6
iso8601==0.1.12
mccabe==0.6.1
numpy==1.13.3
oauthlib==2.0.6
pandas==0.21.0
plaid-python==2.1.1
promise==2.1
psycopg2==2.7.3.2
py==1.5.2
pycodestyle==2.3.1
pycparser==2.18
pyflakes==1.6.0
PyJWT==1.5.3
pyOpenSSL==17.5.0
pytest==3.3.1
pytest-django==3.1.2
python-dateutil==2.6.1
pytest-selenium==1.11.3
python3-openid==3.1.0
pytz==2017.3
raven==6.3.0
requests==2.18.4
requests-oauthlib==0.8.0
selenium==3.8.0
singledispatch==3.4.0.3
six==1.11.0
snapshottest==0.5.0
social-auth-app-django==2.0.0
social-auth-core==1.5.0
termcolor==1.1.0
typing==3.6.2
ua-parser==0.7.3
urllib3==1.22
whitenoise==3.3.1
yahoo-historical==0.1.4
